improve methods security

+++ manager_app/src/main/java/com/example/managerapp/service/ImageService.java
@@ -1,7 +1,10 @@
 package com.example.managerapp.service;
 
+import com.example.managerapp.entity.MyUser;
 import com.example.managerapp.entity.enums.MinIO_bucket;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -18,26 +21,28 @@ public class ImageService {
 
 
 
-    public void saveUserAvatar(MultipartFile file, Long userId) throws IOException {
+    public void saveUserAvatar(MultipartFile file, OAuth2AuthenticationToken authenticationToken) throws IOException {
+
+
 
         String newFileName =  minioService.saveFile(file, MinIO_bucket.users);
 
-        userService.getMyUserFromPostgres(userId).ifPresent(myUser -> {
+        MyUser myUser = userService.getMyUserFromPostgres(authenticationToken);
             if(myUser.getAvatarFileName()!=null){
                 minioService.deleteFile(myUser.getAvatarFileName(), MinIO_bucket.users);
             }
             myUser.setAvatarFileName(newFileName);
             userService.saveMyUserToPostgres(myUser);
-        });
     }
 
-    public void deleteUserAvatar(Long userId) {
+    public void deleteUserAvatar(OAuth2AuthenticationToken authenticationToken) {
 
-        userService.getMyUserFromPostgres(userId).ifPresent(myUser -> {
+        MyUser myUser = userService.getMyUserFromPostgres(authenticationToken);
+            if(myUser.getAvatarFileName()!=null){
             minioService.deleteFile(myUser.getAvatarFileName(),MinIO_bucket.users);
             myUser.setAvatarFileName(null);
             userService.saveMyUserToPostgres(myUser);
-        });
+            }
     }
